Added Airline, nvim terminal maps, and snippets
- Airline is just about as fast as lightline and the buffer tabs are
just too useful to pass up

- Map window movements to CTRL-h/j/k/l for easy of use in neovim,
because now windows navigation is uniform accross neovim and the
terminal emulator

- Implement two different snippet strageties that are native to vim. one
uses abbreviations, so one can type teh abbreviation and space in order
for the snippet to expand. However, this does not allow for more complex
movements inside the snippet. The second stragegy is to implement insert
mappings that are triggered with a different leader than space. this is
becasuse space makes vim wait on every space to look for the second
chord, which results in a visually laggy experience. The unique insert
escape chosen is '\' because it is rarely used and was the original
escape character in vim.