Add buffertab and copy to clipboard functionality
I intergrated the bufferline plugin into lightline, and changed the
buffer next and buffer previous commands back to their previous values.

I also added a mapping for Ctrl-c which will yank to the system
clipboard. An example of usage would be "<C-c>$" to yank from the
current position to the end of the line.